Fix i32/u32 overflow in pixel indexing above 2.1 billion pixels
Exporting a 50000x50000 image with a diagonal pixel sort panicked:

  src\modifiers\pixel_sort.rs:201
  index out of bounds: the len is 2500000000
  but the index is 18446744071562084318

`for_each_diagonal_line` computed `(y * w + x) as usize` with i32
operands. A 50000x50000 image has 2.5e9 pixels, so `y * w` passes
i32::MAX at y = 42949, wraps negative, and casts to a ~1.8e19 usize.
The reported index decodes exactly to y=42949, x=49998, which confirms
the mechanism rather than inferring it.

Computing the index in usize fixes it. Verified end to end: a diagonal
sort at 50000x50000 now completes in ~114s instead of panicking.

Two more instances of the same defect, found by sweeping for the
pattern rather than waiting for reports:

- image_data.rs XCF compositing: `(py * width + px) as usize` with u32
  operands overflows past ~4.29e9 pixels, wrapping to a small index and
  silently corrupting the composite rather than panicking.
- passes/drawing.rs upload: `((y0 * w + x0) * 4) as usize` with u32
  operands, same exposure on a very large canvas.

Added `diagonal_indices_survive_more_than_i32_max_pixels`, which walks
the index arithmetic at 50000x50000 without allocating any pixels, so it
runs in under 4s. Confirmed it fails with the old expression (reports a
~1.8e19 index) and passes with the fix.

Not the cause, but worth recording: this path also holds three
full-resolution buffers alive at once (source, processed, output), which
is 30 GB at this size. That is a separate design problem -- export does
not tile, unlike the GPU preview -- and is not addressed here.

This bug predates the render-refactor branch; it is not a regression
introduced by it.

+++ src/wgpu/media/image_data.rs
@@ -670,8 +670,11 @@ impl ImageData {
             let buf = layer.raw_rgba_buffer();
             for py in 0..lh.min(height) {
                 for px in 0..lw.min(width) {
-                    let src = buf[(py * lw + px) as usize].0;
-                    let dst = (py * width + px) as usize * 4;
+                    // Widen before multiplying: these products exceed u32 once
+                    // the layer passes ~4.29 billion pixels, wrapping to a
+                    // small index and silently corrupting the composite.
+                    let src = buf[py as usize * lw as usize + px as usize].0;
+                    let dst = (py as usize * width as usize + px as usize) * 4;
                     let sa = src[3] as u32;
                     let da = canvas[dst + 3] as u32;
                     let out_a = sa + da * (255 - sa) / 255;
