+++ src/wgpu/passes/drawing.rs
@@ -64,8 +64,10 @@ impl DrawingLayer {
             return None;
         }
         let w = self.cache.w;
-        let offset = ((y0 * w + x0) * 4) as usize;
-        let len = (((y1 - 1 - y0) * w + (x1 - x0)) * 4) as usize;
+        // Widen before multiplying: on a very large canvas these products
+        // overflow u32 and produce a wrapped slice range.
+        let offset = (y0 as usize * w as usize + x0 as usize) * 4;
+        let len = ((y1 - 1 - y0) as usize * w as usize + (x1 - x0) as usize) * 4;
         queue.write_texture(
             iced::wgpu::TexelCopyTextureInfo {
                 texture: &self.texture,
